requirements.txt
faiss-cpu>=1.13.1
flask>=3.1.2
flask-cors>=6.0.2
groq>=1.0.0
numpy>=2.3.5
pypdf2>=3.0.1
python-docx>=1.2.0
python-dotenv>=1.2.1
